% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{AbstractColumnDefinition}
\alias{AbstractColumnDefinition}
\title{AbstractColumnDefinition}
\description{
Abstract class to store columns
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{col_type}}{(\code{character(1)})\cr
Type of the column, Read-only}

\item{\code{name}}{(\code{character(1)})\cr
Name of the column, Read-only}

\item{\code{description}}{(\code{character(1)})\cr
Description of the column, Read-only}

\item{\code{dtype}}{(\code{character(1)})\cr
Data type of the column, Read-only}

\item{\code{required}}{(\code{logical})\cr
If the column is required, Read-only}

\item{\code{default}}{(\code{character(1)})\cr
The default of the column, Read-only}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractColumnDefinition-new}{\code{AbstractColumnDefinition$new()}}
\item \href{#method-AbstractColumnDefinition-is_allowed}{\code{AbstractColumnDefinition$is_allowed()}}
\item \href{#method-AbstractColumnDefinition-clone}{\code{AbstractColumnDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractColumnDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractColumnDefinition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the AbstractColumnDefinition class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractColumnDefinition$new(col_type, name, description, dtype, required)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col_type}}{(\code{data.frame})\cr
Type of the column.}

\item{\code{name}}{(\code{character(1)})\cr
Name of the column.}

\item{\code{description}}{(\code{character(1)})\cr
Description of the column.}

\item{\code{dtype}}{(\verb{Data type})\cr
Data type of the column.}

\item{\code{required}}{(\code{Logical})\cr
Bool that specifies if the column is required.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractColumnDefinition-is_allowed"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractColumnDefinition-is_allowed}{}}}
\subsection{Method \code{is_allowed()}}{
Tests if cell is allowed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractColumnDefinition$is_allowed(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell to test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractColumnDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractColumnDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractColumnDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
