% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{AbstractFieldDefinition}
\alias{AbstractFieldDefinition}
\title{AbstractFieldDefinition}
\description{
Abstract class to store fields
}
\examples{

## ------------------------------------------------
## Method `AbstractFieldDefinition$select_and_expand`
## ------------------------------------------------

# Example usage:
# select_and_expand(df, "col_id", field_vals = NULL)

}
\section{Super class}{
\code{\link[posted:AbstractColumnDefinition]{posted::AbstractColumnDefinition}} -> \code{AbstractFieldDefinition}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{field_type}}{(\code{character(1)}):
Type of the field, Read-only}

\item{\code{coded}}{(\code{logical}):
If the field is coded, Read-only}

\item{\code{codes}}{:
Codes of the field, Read-only}

\item{\code{default}}{(\code{character(1)}):
Default of the field, Read-only}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractFieldDefinition-new}{\code{AbstractFieldDefinition$new()}}
\item \href{#method-AbstractFieldDefinition-is_allowed}{\code{AbstractFieldDefinition$is_allowed()}}
\item \href{#method-AbstractFieldDefinition-select_and_expand}{\code{AbstractFieldDefinition$select_and_expand()}}
\item \href{#method-AbstractFieldDefinition-clone}{\code{AbstractFieldDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractFieldDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractFieldDefinition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the AbstractFieldDefinition Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractFieldDefinition$new(
  field_type,
  name,
  description,
  dtype,
  coded,
  codes = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field_type}}{Type of the field}

\item{\code{name}}{Name of the field}

\item{\code{description}}{Description of the field}

\item{\code{dtype}}{Data type of the field}

\item{\code{coded}}{If the field is coded}

\item{\code{codes}}{Optional codes for the field (default: NULL)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractFieldDefinition-is_allowed"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractFieldDefinition-is_allowed}{}}}
\subsection{Method \code{is_allowed()}}{
Tests if cell is allowed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractFieldDefinition$is_allowed(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell to test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractFieldDefinition-select_and_expand"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractFieldDefinition-select_and_expand}{}}}
\subsection{Method \code{select_and_expand()}}{
Select and expand fields which are valid for multiple periods or other field vals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractFieldDefinition$select_and_expand(df, col_id, field_vals = NA, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{DataFrame where fields should be selected and expanded}

\item{\code{col_id}}{col_id of the column to be selected and expanded}

\item{\code{field_vals}}{NULL or list of field_vals to select and expand}

\item{\code{...}}{Additional keyword arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DataFrame where fields are selected and expanded
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example usage:
# select_and_expand(df, "col_id", field_vals = NULL)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractFieldDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractFieldDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractFieldDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
