% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking.R
\name{apply_cond}
\alias{apply_cond}
\title{apply_cond}
\usage{
apply_cond(df, cond)
}
\arguments{
\item{df}{DataFrame. A pandas DataFrame containing the data on which the condition will be applied.}

\item{cond}{MaskCondition. The condition to be applied on the dataframe. Can be either a string, a dictionary, or a
callable function.}
}
\value{
DataFrame. Dataframe evaluated at the mask condition.
}
\description{
Takes a pandas DataFrame and a condition, which can be a string, dictionary,
or callable, and applies the condition to the DataFrame using \code{eval} or \code{apply}
accordingly.
}
