% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_config.R
\docType{data}
\name{baseFormat}
\alias{baseFormat}
\title{Entry type specifications and their default data types.}
\format{
An object of class \code{list} of length 13.
}
\usage{
(NOT EXPORTED)
}
\description{
Entry type specifications and their default data types.
}
\seealso{
Uses \link{readYAMLDataFile}.
}
\keyword{datasets}
