% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_units.R
\docType{data}
\name{cachedUnits}
\alias{cachedUnits}
\alias{cachedUnitsPath}
\title{The cached conversion factors between units.}
\format{
An object of class \code{character} of length 1.
}
\usage{
(NOT EXPORTED)
}
\description{
The cached conversion factors between units.
}
\seealso{
Uses \link{pathOfDataFile}.
}
\keyword{datasets}
