% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noslag.R
\name{collect_files}
\alias{collect_files}
\title{collect_files}
\usage{
collect_files(parent_variable, include_databases = NULL)
}
\arguments{
\item{parent_variable}{Character. Variable to collect files on.}

\item{include_databases}{Optional list\link{Character}. List of Database IDs to collect files from.}
}
\value{
List of tuples. List of tuples containing the parent variable and the
database ID for each file found in the specified directories.
}
\description{
Takes a parent variable and optional list of databases to include,
checks for their existence, and collects files and directories based on the parent variable.
}
\examples{
# Example usage:
collect_files("variable_name", c("db1", "db2"))

}
