% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noslag.R
\name{combine_units}
\alias{combine_units}
\title{combine_units}
\usage{
combine_units(numerator, denominator)
}
\arguments{
\item{numerator}{Character. Numerator of the fraction.}

\item{denominator}{Character. Denominator of the fraction.}
}
\value{
Character. Updated unit string after simplification.
}
\description{
Combine fraction of two units into an updated unit string
}
\examples{
# Example usage:
combine_units("m", "s")

}
