% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{CommentDefinition}
\alias{CommentDefinition}
\title{CommentDefinition}
\description{
Class to store comment columns
}
\section{Super class}{
\code{\link[posted:AbstractColumnDefinition]{posted::AbstractColumnDefinition}} -> \code{CommentDefinition}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CommentDefinition-new}{\code{CommentDefinition$new()}}
\item \href{#method-CommentDefinition-is_allowed}{\code{CommentDefinition$is_allowed()}}
\item \href{#method-CommentDefinition-clone}{\code{CommentDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-CommentDefinition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the CommentDefinition Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentDefinition$new(name, description, required)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character. Name of the column.}

\item{\code{description}}{Character. Description of the column.}

\item{\code{required}}{Logical. Bool that specifies if the column is required.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentDefinition-is_allowed"></a>}}
\if{latex}{\out{\hypertarget{method-CommentDefinition-is_allowed}{}}}
\subsection{Method \code{is_allowed()}}{
Tests if cell is allowed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentDefinition$is_allowed(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell to test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CommentDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
