% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_units.R
\name{convUnit}
\alias{convUnit}
\title{Get conversion factor between units.}
\usage{
convUnit(unit_from, unit_to, flow_type = NULL)
}
\arguments{
\item{unit_from}{The unit to convert from.}

\item{unit_to}{The unit to convert to.}

\item{flow_type}{The flow type.}
}
\value{
The conversion factor.
}
\description{
Get conversion factor between units.
}
\examples{
convUnit("MWh;LHV", "m³;norm")
}
