% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_units.R
\name{convUnitDF}
\alias{convUnitDF}
\title{Convert a dataframe column from one unit to another.}
\usage{
convUnitDF(df, unit_from_col, unit_to_col, flow_type = NULL)
}
\arguments{
\item{df}{The dataframe.}

\item{unit_from_col}{The column with the unit to convert from.}

\item{unit_to_col}{The column with the unit to convert to.}

\item{flow_type}{The flow type.}
}
\value{
A one column data frame that contains the conversion factors
}
\description{
Convert a dataframe column from one unit to another.
}
\examples{
convUnitDF(data.frame(unit_from=c("MWh;LHV", "MWh;LHV"), unit_to=c("m³;norm", "m³;norm")))
}
