% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{CustomFieldDefinition}
\alias{CustomFieldDefinition}
\title{CustomFieldDefinition}
\description{
Class to store Custom fields
}
\section{Super classes}{
\code{\link[posted:AbstractColumnDefinition]{posted::AbstractColumnDefinition}} -> \code{\link[posted:AbstractFieldDefinition]{posted::AbstractFieldDefinition}} -> \code{CustomFieldDefinition}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{field_specs}}{Specs of the field}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CustomFieldDefinition-new}{\code{CustomFieldDefinition$new()}}
\item \href{#method-CustomFieldDefinition-clone}{\code{CustomFieldDefinition$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="posted" data-topic="AbstractFieldDefinition" data-id="is_allowed"><a href='../../posted/html/AbstractFieldDefinition.html#method-AbstractFieldDefinition-is_allowed'><code>posted::AbstractFieldDefinition$is_allowed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="posted" data-topic="AbstractFieldDefinition" data-id="select_and_expand"><a href='../../posted/html/AbstractFieldDefinition.html#method-AbstractFieldDefinition-select_and_expand'><code>posted::AbstractFieldDefinition$select_and_expand()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomFieldDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-CustomFieldDefinition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the CustomFieldDefinition class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomFieldDefinition$new(field_specs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field_specs}}{Specs of the custom field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CustomFieldDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CustomFieldDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CustomFieldDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
