% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noslag.R
\name{DataSet}
\alias{DataSet}
\title{DataSet}
\description{
This class provides methods to store, normalize, select, and aggregate DataSets.
}
\examples{

## ------------------------------------------------
## Method `DataSet$normalise`
## ------------------------------------------------

# Example usage:
dataset$normalize(override = list("variable1" = "value1"), inplace = FALSE)


## ------------------------------------------------
## Method `DataSet$select`
## ------------------------------------------------

# Example usage:
dataset$select(override = list("variable1" = "value1"), drop_singular_fields = TRUE, extrapolate_period = FALSE, field1 = "value1")


## ------------------------------------------------
## Method `DataSet$aggregate`
## ------------------------------------------------

# Example usage:
dataset$aggregate(override = list("variable1" = "value1"), drop_singular_fields = TRUE, extrapolate_period = FALSE, agg = "field", masks = list(mask1, mask2), masks_database = TRUE)

}
\section{Super class}{
\code{\link[posted:TEBase]{posted::TEBase}} -> \code{DataSet}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{access dataframe}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSet-new}{\code{DataSet$new()}}
\item \href{#method-DataSet-normalise}{\code{DataSet$normalise()}}
\item \href{#method-DataSet-select}{\code{DataSet$select()}}
\item \href{#method-DataSet-aggregate}{\code{DataSet$aggregate()}}
\item \href{#method-DataSet-clone}{\code{DataSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSet-new}{}}}
\subsection{Method \code{new()}}{
Create new instance of the DataSet class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSet$new(
  parent_variable,
  include_databases = NULL,
  file_paths = NULL,
  check_inconsistencies = FALSE,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent_variable}}{Character. Variable to collect Data on.}

\item{\code{include_databases}}{Optional list\link{Character} | tuple\link{Character}, optional. Databases to load from.}

\item{\code{file_paths}}{Optional list\link{Character}, optional. Paths to load data from.}

\item{\code{check_inconsistencies}}{Logical, optional. Whether to check for inconsistencies.}

\item{\code{data}}{Optional DataFrame, optional. Specific data to include in the dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSet-normalise"></a>}}
\if{latex}{\out{\hypertarget{method-DataSet-normalise}{}}}
\subsection{Method \code{normalise()}}{
Normalize data: default reference units, reference value equal to 1.0, default reported units
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSet$normalise(override = NULL, inplace = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{override}}{Optional list\link{Character}. Dictionary with key, value pairs of variables to override.}

\item{\code{inplace}}{Logical, optional. Whether to do the normalization in place.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DataFrame. If \code{inplace} is \code{FALSE}, returns normalized dataframe.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example usage:
dataset$normalize(override = list("variable1" = "value1"), inplace = FALSE)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSet-select"></a>}}
\if{latex}{\out{\hypertarget{method-DataSet-select}{}}}
\subsection{Method \code{select()}}{
Select desired data from the dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSet$select(
  override = NULL,
  drop_singular_fields = TRUE,
  extrapolate_period = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{override}}{Optional list\link{Character}. Dictionary with key, value pairs of variables to override.}

\item{\code{drop_singular_fields}}{Logical, optional. If \code{TRUE}, drop custom fields with only one value.}

\item{\code{extrapolate_period}}{Logical, optional. If \code{TRUE}, extrapolate values if no value for this period is given.}

\item{\code{...}}{IDs of values to select.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DataFrame. DataFrame with selected values.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example usage:
dataset$select(override = list("variable1" = "value1"), drop_singular_fields = TRUE, extrapolate_period = FALSE, field1 = "value1")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSet-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-DataSet-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
Aggregates data based on specified parameters, applies masks,
and cleans up the resulting DataFrame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSet$aggregate(
  override = NULL,
  drop_singular_fields = TRUE,
  extrapolate_period = TRUE,
  agg = NULL,
  masks = NULL,
  masks_database = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{override}}{Optional list\link{Character}. Dictionary with key, value pairs of variables to override.}

\item{\code{drop_singular_fields}}{Logical, optional. If \code{TRUE}, drop custom fields with only one value.}

\item{\code{extrapolate_period}}{Logical, optional. If \code{TRUE}, extrapolate values if no value for this period is given.}

\item{\code{agg}}{Optional Character | list\link{Character} | tuple\link{Character}. Specifies which fields to aggregate over.}

\item{\code{masks}}{Optional list\link{Mask}. Specifies a list of Mask objects that will be applied to the data during aggregation. These masks can be used to filter or weight the data based on certain conditions defined in the Mask objects.}

\item{\code{masks_database}}{Logical, optional. Determines whether to include masks from databases in the aggregation process. If \code{TRUE}, masks from databases will be included along with any masks provided as function arguments. If \code{FALSE}, only the masks provided as function arguments will be applied.}

\item{\code{...}}{additional field vals}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DataFrame. The \code{aggregate} method returns a pandas DataFrame that has been cleaned up and aggregated based on the specified parameters and input data. The method performs aggregation over component fields and case fields, applies weights based on masks, drops rows with NaN weights, aggregates with weights, inserts reference variables, sorts columns and rows, rounds values, and inserts units before returning the final cleaned and aggregated DataFrame.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example usage:
dataset$aggregate(override = list("variable1" = "value1"), drop_singular_fields = TRUE, extrapolate_period = FALSE, agg = "field", masks = list(mask1, mask2), masks_database = TRUE)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
