% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_config.R
\docType{data}
\name{defaultDtypes}
\alias{defaultDtypes}
\alias{mapColnamesDtypes}
\title{Default data type for each entry type.}
\format{
An object of class \code{list} of length 13.
}
\usage{
(NOT EXPORTED)
}
\description{
Default data type for each entry type.
}
\keyword{datasets}
