% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_config.R
\docType{data}
\name{defaultMasks}
\alias{defaultMasks}
\title{Default masks that are applied to the flow types.}
\format{
An object of class \code{list} of length 1.
}
\usage{
defaultMasks
}
\description{
Default masks that are applied to the flow types.
}
\keyword{datasets}
