% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_config.R
\docType{data}
\name{defaultUnits}
\alias{defaultUnits}
\title{The default units for some dimensions (e.g. time, currency).}
\format{
An object of class \code{list} of length 2.
}
\usage{
defaultUnits
}
\description{
The default units for some dimensions (e.g. time, currency).
}
\keyword{datasets}
