% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_config.R
\docType{data}
\name{dtypeMapping}
\alias{dtypeMapping}
\title{Mapping of pandas dtypes to R dataframe types.}
\format{
An object of class \code{list} of length 3.
}
\usage{
(NOT EXPORTED)
}
\description{
Mapping of pandas dtypes to R dataframe types.
}
\keyword{datasets}
