% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_config.R
\docType{data}
\name{flowTypes}
\alias{flowTypes}
\title{Flow type information (e.g. default unit, energy content, density).}
\format{
An object of class \code{list} of length 26.
}
\usage{
(NOT EXPORTED)
}
\description{
Flow type information (e.g. default unit, energy content, density).
}
\seealso{
Uses \link{readCSVDataFile}.
}
\keyword{datasets}
