% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{is_float}
\alias{is_float}
\title{is_float}
\usage{
is_float(string)
}
\arguments{
\item{string}{Character. String to check.}
}
\value{
Logical. \code{TRUE} if conversion was successful, \code{FALSE} if not.
}
\description{
Checks if a given string can be converted to a floating-point number in Python.
}
\examples{
# Example usage:
is_numeric("3.14")

}
