% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking.R
\name{Mask}
\alias{Mask}
\title{Mask}
\description{
Class to define masks with conditions and weights to apply to DataFiles
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Mask-new}{\code{Mask$new()}}
\item \href{#method-Mask-matches}{\code{Mask$matches()}}
\item \href{#method-Mask-get_weights}{\code{Mask$get_weights()}}
\item \href{#method-Mask-clone}{\code{Mask$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mask-new"></a>}}
\if{latex}{\out{\hypertarget{method-Mask-new}{}}}
\subsection{Method \code{new()}}{
Create a new mask object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mask$new(where = NULL, use = NULL, weight = NULL, other = NaN, comment = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{where}}{MaskCondition | list\link{MaskCondition}, optional. Where the mask should be applied.}

\item{\code{use}}{MaskCondition | list\link{MaskCondition}, optional. Condition on where to use the masks.}

\item{\code{weight}}{Numeric | Character | list\link{Numeric | Character}, optional. Weights to apply.}

\item{\code{other}}{Numeric, optional.}

\item{\code{comment}}{Character, optional. Comment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mask-matches"></a>}}
\if{latex}{\out{\hypertarget{method-Mask-matches}{}}}
\subsection{Method \code{matches()}}{
Check if a mask matches a dataframe by verifying if all 'where' conditions match across all rows.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mask$matches(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{DataFrame. Dataframe to check for matches.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical. If the mask matches the dataframe.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mask-get_weights"></a>}}
\if{latex}{\out{\hypertarget{method-Mask-get_weights}{}}}
\subsection{Method \code{get_weights()}}{
Apply weights to the dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mask$get_weights(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{(\code{Dataframe}): Dataframe to apply weights on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dataframe. Dataframe with applied weights
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Mask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
