% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noslag.R
\name{normalise_units}
\alias{normalise_units}
\title{normalise_units}
\usage{
normalise_units(df, level, var_units, var_flow_ids)
}
\arguments{
\item{df}{DataFrame. Dataframe to be normalized.}

\item{level}{Character. Specifies whether the data should be normalized on the reported or reference values. Possible values are 'reported' or 'reference'.}

\item{var_units}{List. Dictionary that maps a combination of parent variable and variable to its corresponding unit. The keys in the dictionary are in the format "{parent_variable}|{variable}", and the values are the units associated with that variable.}

\item{var_flow_ids}{List. Dictionary that maps a combination of parent variable and variable to a specific flow ID. This flow ID is used for unit conversion in the \code{normalize_units} function.}
}
\value{
DataFrame. Normalized dataframe.
}
\description{
Takes a DataFrame with reported or reference data, along with
dictionaries mapping variable units and flow IDs, and normalizes the units of the variables in the
DataFrame based on the provided mappings.
}
\examples{
# Example usage:
normalize_dataframe(df, "reported", var_units, var_flow_ids)

}
