% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noslag.R
\name{normalise_values}
\alias{normalise_values}
\title{normalise_values}
\usage{
normalise_values(df)
}
\arguments{
\item{df}{DataFrame. Dataframe to be normalized.}
}
\value{
DataFrame. Returns a modified DataFrame where the 'value' column has been
divided by the 'reference_value' column (or 1.0 if 'reference_value' is null), the 'uncertainty'
column has been divided by the 'reference_value' column, and the 'reference_value' column has been
replaced with 1.0 if it was not null.
}
\description{
Takes a DataFrame as input, normalizes the 'value' and 'uncertainty'
columns by the reference value, and updates the 'reference_value' column accordingly.
}
\examples{
# Example usage:
normalized_df <- normalize_values(df)

}
