% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{pathOfDataFile}
\alias{pathOfDataFile}
\title{Path of a data file}
\usage{
(NOT EXPORTED)
}
\arguments{
\item{fname}{The file name.}
}
\description{
This function returns the path of a file in the data directory.
}
