% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{pathOfFile}
\alias{pathOfFile}
\title{Path of a file in a given directory}
\usage{
(NOT EXPORTED)
}
\arguments{
\item{dname}{The directory name.}

\item{fname}{The file name.}
}
\description{
This function returns the path of a file in a given directory.
}
