% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{pathOfOutputFile}
\alias{pathOfOutputFile}
\title{Path of an output file}
\usage{
(NOT EXPORTED)
}
\arguments{
\item{fname}{The file name.}
}
\description{
This function returns the path of a file in the output directory.
}
