% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{pathOfTEDFile}
\alias{pathOfTEDFile}
\title{Path of a TED file}
\usage{
(NOT EXPORTED)
}
\arguments{
\item{tid}{The technology ID.}
}
\description{
This function returns the path of a TED data file in the TED directory.
}
