% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{PeriodFieldDefinition}
\alias{PeriodFieldDefinition}
\title{PeriodFieldDefinition}
\description{
Class to store Period fields
}
\section{Super classes}{
\code{\link[posted:AbstractColumnDefinition]{posted::AbstractColumnDefinition}} -> \code{\link[posted:AbstractFieldDefinition]{posted::AbstractFieldDefinition}} -> \code{PeriodFieldDefinition}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PeriodFieldDefinition-new}{\code{PeriodFieldDefinition$new()}}
\item \href{#method-PeriodFieldDefinition-is_allowed}{\code{PeriodFieldDefinition$is_allowed()}}
\item \href{#method-PeriodFieldDefinition-clone}{\code{PeriodFieldDefinition$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="posted" data-topic="AbstractFieldDefinition" data-id="select_and_expand"><a href='../../posted/html/AbstractFieldDefinition.html#method-AbstractFieldDefinition-select_and_expand'><code>posted::AbstractFieldDefinition$select_and_expand()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodFieldDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodFieldDefinition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the PeriodFieldDefinition Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodFieldDefinition$new(name, description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character. Name of the field.}

\item{\code{description}}{Character. Description of the field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodFieldDefinition-is_allowed"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodFieldDefinition-is_allowed}{}}}
\subsection{Method \code{is_allowed()}}{
Tests if cell is allowed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodFieldDefinition$is_allowed(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell to test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PeriodFieldDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PeriodFieldDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PeriodFieldDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
