% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{read_definitions}
\alias{read_definitions}
\title{read_definitions}
\usage{
read_definitions(definitions_dir, flows, techs)
}
\arguments{
\item{definitions_dir}{Character. Path leading to the definitions.}

\item{flows}{List. Dictionary containing the different flow types. Each key represents a flow type, the corresponding
value is a dictionary containing key value pairs of attributes like density, energy content and their
values.}

\item{techs}{List. Dictionary containing information about different technologies. Each key in the
dictionary represents a unique technology ID, and the corresponding value is a dictionary containing
various specifications for that technology, like 'description', 'class', 'primary output' etc.}
}
\value{
List. Dictionary containing the definitions after processing and replacing tags and tokens.
}
\description{
Reads YAML files from definitions directory, extracts tags, inserts tags into
definitions, replaces tokens in definitions, and returns the updated definitions.
}
