% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking.R
\name{read_masks}
\alias{read_masks}
\title{read_masks}
\usage{
read_masks(variable)
}
\arguments{
\item{variable}{Character. Variable to be read.}
}
\value{
List. List with masks for the variable.
}
\description{
Reads YAML files containing mask specifications from multiple databases and returns a list of Mask objects.
}
