% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_read_config.R
\name{readCSVDataFile}
\alias{readCSVDataFile}
\title{Read a CSV data file.}
\usage{
(NOT EXPORTED)
}
\arguments{
\item{fname}{The relative file path.}
}
\description{
This function reads in a csv file specified by the relative file path.
}
\seealso{
Uses \link{pathOfDataFile}.
}
