% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_read_config.R
\name{readYAMLDataFile}
\alias{readYAMLDataFile}
\title{Read a YAML data file.}
\usage{
(NOT EXPORTED)
}
\arguments{
\item{fname}{The relative file path (without the file extension).}
}
\description{
This function reads in a yaml file specified by the relative file path (without the file extension).
}
\seealso{
Uses \link{pathOfDataFile}.
}
