% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{replace_tags}
\alias{replace_tags}
\title{replace_tags}
\usage{
replace_tags(definitions, tag, items)
}
\arguments{
\item{definitions}{List. Dictionary containing the definitions, where the tags should be replaced by the items.}

\item{tag}{Character. String to identify where replacements should be made in the definitions. Specifies
the placeholder that needs to be replaced with actual values from the \code{items} dictionary.}

\item{items}{List. Dictionary containing the items from which to replace the definitions.}
}
\value{
List. Dictionary containing the definitions with replacements based on the provided tag and items.
}
\description{
Replaces specified tags in dictionary keys and values with corresponding
items from another dictionary.
}
