% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{SourceFieldDefinition}
\alias{SourceFieldDefinition}
\title{SourceFieldDefinition}
\description{
Class to store Source fields
}
\section{Super classes}{
\code{\link[posted:AbstractColumnDefinition]{posted::AbstractColumnDefinition}} -> \code{\link[posted:AbstractFieldDefinition]{posted::AbstractFieldDefinition}} -> \code{SourceFieldDefinition}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SourceFieldDefinition-new}{\code{SourceFieldDefinition$new()}}
\item \href{#method-SourceFieldDefinition-clone}{\code{SourceFieldDefinition$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="posted" data-topic="AbstractFieldDefinition" data-id="is_allowed"><a href='../../posted/html/AbstractFieldDefinition.html#method-AbstractFieldDefinition-is_allowed'><code>posted::AbstractFieldDefinition$is_allowed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="posted" data-topic="AbstractFieldDefinition" data-id="select_and_expand"><a href='../../posted/html/AbstractFieldDefinition.html#method-AbstractFieldDefinition-select_and_expand'><code>posted::AbstractFieldDefinition$select_and_expand()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceFieldDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-SourceFieldDefinition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the SourceFieldDefinition class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceFieldDefinition$new(name, description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character. Name of the field.}

\item{\code{description}}{Character. Description of the field.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceFieldDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SourceFieldDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceFieldDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
