% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tedf.R
\name{TEBase}
\alias{TEBase}
\title{TEBase}
\description{
This is the base class for technoeconomic data.
}
\examples{
# Example usage:
base_technoeconomic_data <- TEBase$new("variable_name")

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parent_variable}}{Get parent variable, Read-only}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TEBase-new}{\code{TEBase$new()}}
\item \href{#method-TEBase-clone}{\code{TEBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-TEBase-new}{}}}
\subsection{Method \code{new()}}{
Create new instance of TEBase class. Set parent variable and technology specifications (var_specs) from input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEBase$new(parent_variable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent_variable}}{(\code{character})
Name of the parent variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TEBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
