% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_config.R
\docType{data}
\name{techClasses}
\alias{techClasses}
\title{Technology classes (e.g. storage, conversion and transport) specifications and their entry types.}
\format{
An object of class \code{list} of length 3.
}
\usage{
(NOT EXPORTED)
}
\description{
Technology classes (e.g. storage, conversion and transport) specifications and their entry types.
}
\seealso{
Uses \link{readYAMLDataFile}.
}
\keyword{datasets}
