% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_config.R
\docType{data}
\name{techs}
\alias{techs}
\title{Technology specifications (e.g. name, desc, sector, case_fields, ...).}
\format{
An object of class \code{list} of length 19.
}
\usage{
(NOT EXPORTED)
}
\description{
Technology specifications (e.g. name, desc, sector, case_fields, ...).
}
\seealso{
Uses \link{readYAMLDataFile}.
}
\keyword{datasets}
