% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_TEDataFile.R
\name{TEDataFile}
\alias{TEDataFile}
\title{The base class for all TE data file classes.}
\usage{
(NOT EXPORTED)
}
\description{
This abstract class defines the basic structure of a TE data file class.
}
\section{Super class}{
\code{posted::TEBase} -> \code{TEDataFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TEDataFile-new}{\code{TEDataFile$new()}}
\item \href{#method-TEDataFile-load}{\code{TEDataFile$load()}}
\item \href{#method-TEDataFile-read}{\code{TEDataFile$read()}}
\item \href{#method-TEDataFile-write}{\code{TEDataFile$write()}}
\item \href{#method-TEDataFile-data}{\code{TEDataFile$data()}}
\item \href{#method-TEDataFile-getInconsistencies}{\code{TEDataFile$getInconsistencies()}}
\item \href{#method-TEDataFile-check}{\code{TEDataFile$check()}}
\item \href{#method-TEDataFile-checkRow}{\code{TEDataFile$checkRow()}}
\item \href{#method-TEDataFile-clone}{\code{TEDataFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="posted" data-topic="TEBase" data-id="getDataFormat"><a href='../../posted/html/TEBase.html#method-TEBase-getDataFormat'><code>posted::TEBase$getDataFormat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="posted" data-topic="TEBase" data-id="getDtypeMapping"><a href='../../posted/html/TEBase.html#method-TEBase-getDtypeMapping'><code>posted::TEBase$getDtypeMapping()</code></a></span></li>
<li><span class="pkg-link" data-pkg="posted" data-topic="TEBase" data-id="refFlow"><a href='../../posted/html/TEBase.html#method-TEBase-refFlow'><code>posted::TEBase$refFlow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="posted" data-topic="TEBase" data-id="refUnit"><a href='../../posted/html/TEBase.html#method-TEBase-refUnit'><code>posted::TEBase$refUnit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataFile-new"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataFile-new}{}}}
\subsection{Method \code{new()}}{
Create a TEDataFile object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataFile$new(tid, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tid}}{The technology ID.}

\item{\code{path}}{The path to the data file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataFile-load"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataFile-load}{}}}
\subsection{Method \code{load()}}{
Checks if the data has been read in, if not reads the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataFile$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the data from the data file.
Checks if the data file contains no unknown columns.
Inserts missing columns, reorders via reindexing and updates dtypes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataFile$read()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataFile-write}{}}}
\subsection{Method \code{write()}}{
Writes the data back to the data file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataFile$write()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataFile-data"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataFile-data}{}}}
\subsection{Method \code{data()}}{
Get the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataFile$data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The data frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataFile-getInconsistencies"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataFile-getInconsistencies}{}}}
\subsection{Method \code{getInconsistencies()}}{
Get the inconsistencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataFile$getInconsistencies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The inconsistencies.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataFile-check"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataFile-check}{}}}
\subsection{Method \code{check()}}{
Check the data for inconsistencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataFile$check(re = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{re}}{If TRUE, inconsistencies are checked.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataFile-checkRow"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataFile-checkRow}{}}}
\subsection{Method \code{checkRow()}}{
Check the consistency of a single row.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataFile$checkRow(rowID, re = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rowID}}{The row ID.}

\item{\code{re}}{If TRUE, inconsistencies are checked.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
