% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_TEDataSet.R
\name{TEDataSet}
\alias{TEDataSet}
\title{The class that implements a fully usable TE data set.}
\description{
This class implements a fully usable TE data set.
It enables to generate aggregated tables to analize the underlying data.
}
\examples{
elh2 <- TEDataSet$new("elh2")
elh2$data()
}
\section{Super class}{
\code{\link[posted:TEBase]{posted::TEBase}} -> \code{TEDataSet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TEDataSet-new}{\code{TEDataSet$new()}}
\item \href{#method-TEDataSet-data}{\code{TEDataSet$data()}}
\item \href{#method-TEDataSet-getRepUnit}{\code{TEDataSet$getRepUnit()}}
\item \href{#method-TEDataSet-getRefUnit}{\code{TEDataSet$getRefUnit()}}
\item \href{#method-TEDataSet-query}{\code{TEDataSet$query()}}
\item \href{#method-TEDataSet-clone}{\code{TEDataSet$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="posted" data-topic="TEBase" data-id="getDataFormat"><a href='../../posted/html/TEBase.html#method-TEBase-getDataFormat'><code>posted::TEBase$getDataFormat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="posted" data-topic="TEBase" data-id="getDtypeMapping"><a href='../../posted/html/TEBase.html#method-TEBase-getDtypeMapping'><code>posted::TEBase$getDtypeMapping()</code></a></span></li>
<li><span class="pkg-link" data-pkg="posted" data-topic="TEBase" data-id="refFlow"><a href='../../posted/html/TEBase.html#method-TEBase-refFlow'><code>posted::TEBase$refFlow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="posted" data-topic="TEBase" data-id="refUnit"><a href='../../posted/html/TEBase.html#method-TEBase-refUnit'><code>posted::TEBase$refUnit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataSet-new}{}}}
\subsection{Method \code{new()}}{
Create a TEDataSet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataSet$new(
  tid,
  data = NULL,
  load_other = list(),
  load_database = FALSE,
  check_incons = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tid}}{The technology ID.}

\item{\code{data}}{The data frame.}

\item{\code{load_other}}{Other data files to load.}

\item{\code{load_database}}{Whether to load the default data file from the POSTED database.}

\item{\code{check_incons}}{Whether to check for inconsistencies in the data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataSet-data"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataSet-data}{}}}
\subsection{Method \code{data()}}{
Get the data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataSet$data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The data frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataSet-getRepUnit"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataSet-getRepUnit}{}}}
\subsection{Method \code{getRepUnit()}}{
Get the reported unit of an entry type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataSet$getRepUnit(typeid, flowid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{typeid}}{The entry type ID.}

\item{\code{flowid}}{The flow type ID. NULL if not applicable.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataSet-getRefUnit"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataSet-getRefUnit}{}}}
\subsection{Method \code{getRefUnit()}}{
Get the reference unit of an entry type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataSet$getRefUnit(typeid, flowid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{typeid}}{The entry type ID.}

\item{\code{flowid}}{The flow type ID. NULL if not applicable.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataSet-query"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataSet-query}{}}}
\subsection{Method \code{query()}}{
Query the data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataSet$query(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The query parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDataSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TEDataSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDataSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
