% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tedf.R
\name{TEDF}
\alias{TEDF}
\title{TEDF}
\description{
This class is used to store Technoeconomic DataFiles.
}
\note{
This method warns if TEDF is already loaded.
}
\examples{
# Example usage:
tedf <- TEDF$new("variable_name")
tedf$load()
tedf$read("file_path.csv")
tedf$write("output_file_path.csv")
tedf$check()
tedf$check_row()


## ------------------------------------------------
## Method `TEDF$load`
## ------------------------------------------------

# Example usage:
tedf$load()


## ------------------------------------------------
## Method `TEDF$read`
## ------------------------------------------------

# Example usage:
tedf$read()


## ------------------------------------------------
## Method `TEDF$write`
## ------------------------------------------------

# Example usage:
tedf$write()


## ------------------------------------------------
## Method `TEDF$check`
## ------------------------------------------------

# Example usage:
tedf$check(raise_exception = TRUE)

}
\section{Super class}{
\code{\link[posted:TEBase]{posted::TEBase}} -> \code{TEDF}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{file_path}}{: Get or set the file path}

\item{\code{data}}{Get data, i.e. access dataframe. Read-only}

\item{\code{inconsistencies}}{Get the inconsistencies of the TEDF. Read-only
Get inconsistencies}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TEDF-new}{\code{TEDF$new()}}
\item \href{#method-TEDF-load}{\code{TEDF$load()}}
\item \href{#method-TEDF-read}{\code{TEDF$read()}}
\item \href{#method-TEDF-write}{\code{TEDF$write()}}
\item \href{#method-TEDF-check}{\code{TEDF$check()}}
\item \href{#method-TEDF-check_row}{\code{TEDF$check_row()}}
\item \href{#method-TEDF-clone}{\code{TEDF$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDF-new"></a>}}
\if{latex}{\out{\hypertarget{method-TEDF-new}{}}}
\subsection{Method \code{new()}}{
Create new instance of TEDF class. Initialise parent class and object fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDF$new(
  parent_variable,
  database_id = "public",
  file_path = NULL,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent_variable}}{(\code{Character}): Variable from which data should be collected.}

\item{\code{database_id}}{(\code{Character}):, default: "public". Database from which to load data.}

\item{\code{file_path}}{(\code{Path}):, optional. File path from which to load file.}

\item{\code{data}}{(\code{DataFrame}):, optional. Specific Technoeconomic data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDF-load"></a>}}
\if{latex}{\out{\hypertarget{method-TEDF-load}{}}}
\subsection{Method \code{load()}}{
Load TEDataFile (only if it has not been read yet)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDF$load()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TEDF. Returns the TEDF object it is called on.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example usage:
tedf$load()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDF-read"></a>}}
\if{latex}{\out{\hypertarget{method-TEDF-read}{}}}
\subsection{Method \code{read()}}{
This method reads TEDF from a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDF$read()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example usage:
tedf$read()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDF-write"></a>}}
\if{latex}{\out{\hypertarget{method-TEDF-write}{}}}
\subsection{Method \code{write()}}{
write TEDF to CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDF$write()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example usage:
tedf$write()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDF-check"></a>}}
\if{latex}{\out{\hypertarget{method-TEDF-check}{}}}
\subsection{Method \code{check()}}{
Check that TEDF is consistent and add inconsistencies to internal parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDF$check(raise_exception = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raise_exception}}{Logical, default: TRUE. If an exception is to be raised.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example usage:
tedf$check(raise_exception = TRUE)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDF-check_row"></a>}}
\if{latex}{\out{\hypertarget{method-TEDF-check_row}{}}}
\subsection{Method \code{check_row()}}{
checks if row of dataframe has issues - NOT IMPLEMENTED YET
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDF$check_row(row_id, raise_exception = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{row_id}}{Id of the row}

\item{\code{raise_exception}}{(\code{logical}) If exception is to be raised}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDF-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TEDF-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
