% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tedf.R
\name{TEDFInconsistencyException}
\alias{TEDFInconsistencyException}
\title{TEDFInconsistencyException}
\description{
This is a class to store inconsistencies in the TEDFs
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TEDFInconsistencyException-new}{\code{TEDFInconsistencyException$new()}}
\item \href{#method-TEDFInconsistencyException-clone}{\code{TEDFInconsistencyException$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDFInconsistencyException-new"></a>}}
\if{latex}{\out{\hypertarget{method-TEDFInconsistencyException-new}{}}}
\subsection{Method \code{new()}}{
Create instance of TEDFInconsistencyException class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDFInconsistencyException$new(
  message = "Inconsistency detected",
  row_id = NULL,
  col_id = NULL,
  file_path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{(\code{character}) the message of the exception}

\item{\code{row_id}}{Id of the row}

\item{\code{col_id}}{Id of the column}

\item{\code{file_path}}{file path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEDFInconsistencyException-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TEDFInconsistencyException-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEDFInconsistencyException$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
