% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unit_convert}
\alias{unit_convert}
\title{Conversion of units with or withour flow ID}
\usage{
unit_convert(unit_from, unit_to, flow_id = NULL)
}
\arguments{
\item{unit_from}{Character or numeric. Unit to convert from.}

\item{unit_to}{Character or numeric. Unit to convert to.}

\item{flow_id}{Character or NULL. Identifier for the specific flow or process.}
}
\value{
Numeric. Conversion factor between \code{unit_from} and \code{unit_to}.
}
\description{
Converts units with optional flow context handling based on
specified variants and flow ID. The function checks if the input units are not NaN,
then it proceeds to handle different cases based on the presence of a flow context and unit
variants.
}
\examples{
# Example usage:
unit_convert("m", "km", flow_id = NULL)

}
