% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{unit_token_func}
\alias{unit_token_func}
\title{unit_token_func}
\usage{
unit_token_func(unit_component, flows)
}
\arguments{
\item{unit_component}{Character. Specifies the type of unit token to be returned. Possible values are 'full', 'raw', 'variant'.}

\item{flows}{List. Dictionary containing the flows.}
}
\value{
Function. Lambda function that takes a dictionary \code{def_specs} as input. The lambda function
will return different values based on the \code{unit_component} parameter.
}
\description{
Takes a unit component type and a dictionary of flows, and returns a lambda function
that extracts the default unit based on the specified component type from the flow
dictionary.
}
