% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{UnitDefinition}
\alias{UnitDefinition}
\title{UnitDefinition}
\description{
Class to store Unit columns
}
\section{Super class}{
\code{\link[posted:AbstractColumnDefinition]{posted::AbstractColumnDefinition}} -> \code{UnitDefinition}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnitDefinition-new}{\code{UnitDefinition$new()}}
\item \href{#method-UnitDefinition-is_allowed}{\code{UnitDefinition$is_allowed()}}
\item \href{#method-UnitDefinition-clone}{\code{UnitDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnitDefinition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the UnitDefinition class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitDefinition$new(name, description, required)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character. Name of the column.}

\item{\code{description}}{Character. Description of the column.}

\item{\code{required}}{Logical. Bool that specifies if the column is required.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitDefinition-is_allowed"></a>}}
\if{latex}{\out{\hypertarget{method-UnitDefinition-is_allowed}{}}}
\subsection{Method \code{is_allowed()}}{
Tests if cell is allowed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitDefinition$is_allowed(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell to test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnitDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
