% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{ValueDefinition}
\alias{ValueDefinition}
\title{ValueDefinition}
\description{
Class to store Value columns
}
\section{Super class}{
\code{\link[posted:AbstractColumnDefinition]{posted::AbstractColumnDefinition}} -> \code{ValueDefinition}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValueDefinition-new}{\code{ValueDefinition$new()}}
\item \href{#method-ValueDefinition-is_allowed}{\code{ValueDefinition$is_allowed()}}
\item \href{#method-ValueDefinition-clone}{\code{ValueDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValueDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-ValueDefinition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the ValueDefinition class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValueDefinition$new(name, description, required)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character. Name of the column.}

\item{\code{description}}{Character. Description of the column.}

\item{\code{required}}{Logical. Bool that specifies if the column is required.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValueDefinition-is_allowed"></a>}}
\if{latex}{\out{\hypertarget{method-ValueDefinition-is_allowed}{}}}
\subsection{Method \code{is_allowed()}}{
Tests if cell is allowed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValueDefinition$is_allowed(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell to test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValueDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValueDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValueDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
