% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{VariableDefinition}
\alias{VariableDefinition}
\title{VariableDefinition}
\description{
Class to store variable columns
}
\section{Super class}{
\code{\link[posted:AbstractColumnDefinition]{posted::AbstractColumnDefinition}} -> \code{VariableDefinition}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VariableDefinition-new}{\code{VariableDefinition$new()}}
\item \href{#method-VariableDefinition-is_allowed}{\code{VariableDefinition$is_allowed()}}
\item \href{#method-VariableDefinition-clone}{\code{VariableDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-VariableDefinition-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the VariableDefinition class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableDefinition$new(name, description, required)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(\code{Character}):
Name of the column.}

\item{\code{description}}{(\code{Character}):
Description of the column.}

\item{\code{required}}{(\code{Logical}):
Bool that specifies if the column is required.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableDefinition-is_allowed"></a>}}
\if{latex}{\out{\hypertarget{method-VariableDefinition-is_allowed}{}}}
\subsection{Method \code{is_allowed()}}{
Tests if cell is allowed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableDefinition$is_allowed(cell)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell}}{cell to test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VariableDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
