% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_TEBase.R
\name{TEBase}
\alias{TEBase}
\title{The base class for all TE classes.}
\usage{
(NOT EXPORTED)
}
\description{
This abstract class defines the basic structure of a TE class.
This includes tid, data format, technology specifications, and dtype mapping.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TEBase-new}{\code{TEBase$new()}}
\item \href{#method-TEBase-getDataFormat}{\code{TEBase$getDataFormat()}}
\item \href{#method-TEBase-refUnit}{\code{TEBase$refUnit()}}
\item \href{#method-TEBase-refFlow}{\code{TEBase$refFlow()}}
\item \href{#method-TEBase-getDtypeMapping}{\code{TEBase$getDtypeMapping()}}
\item \href{#method-TEBase-clone}{\code{TEBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-TEBase-new}{}}}
\subsection{Method \code{new()}}{
Create a TEBase object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEBase$new(tid = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tid}}{The technology ID.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEBase-getDataFormat"></a>}}
\if{latex}{\out{\hypertarget{method-TEBase-getDataFormat}{}}}
\subsection{Method \code{getDataFormat()}}{
Get the data format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEBase$getDataFormat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of data format specifications.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEBase-refUnit"></a>}}
\if{latex}{\out{\hypertarget{method-TEBase-refUnit}{}}}
\subsection{Method \code{refUnit()}}{
Get the default reference unit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEBase$refUnit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The default reference unit.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEBase-refFlow"></a>}}
\if{latex}{\out{\hypertarget{method-TEBase-refFlow}{}}}
\subsection{Method \code{refFlow()}}{
Get the reference flow.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEBase$refFlow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The reference flow.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEBase-getDtypeMapping"></a>}}
\if{latex}{\out{\hypertarget{method-TEBase-getDtypeMapping}{}}}
\subsection{Method \code{getDtypeMapping()}}{
Get the default data type (data type mapping).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEBase$getDtypeMapping()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The default data type.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TEBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TEBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TEBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
